+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - `multidockerfile version` command ([#1](https://github.com/haines/multidockerfile/pull/1))
+- `multidockerfile join` command ([#2](https://github.com/haines/multidockerfile/pull/2))
 
 
 [Unreleased]: https://github.com/haines/multidockerfile/compare/20586c3eb00aad3dde1ca63eb47dcb14ae6372d5...HEAD